requirements.txt
# Core dependencies with pinned versions for security
opencv-python==4.10.0.84
pyvirtualcam==0.12.0
websocket-client==1.8.0
Pillow==10.4.0
numpy==1.26.4

# Development dependencies
python-dotenv==1.0.1
pytest==8.3.4
black==24.10.0
flake8==7.1.1